Add ETI to LinearBVH.
